requirements.txt
# Python backend (backend_llm) – optional, for uvicorn backend_llm.app:app
fastapi>=0.100.0
uvicorn[standard]>=0.22.0
openai>=1.0.0
python-dotenv>=1.0.0
pydantic>=2.0.0
